Refactor Tar instrument file
Formerly all the notes on the strings of the tar corresponding to each
fret were listed.  That was not scalable, since required listing those
for different strings and would not support different key/tuning.

Now the implementation is based on the `MusicInterval` between
consecutive frets and the open-hand note of each string.

Along the way a bug in the former way of Tar string note generation
was notice and rectified automatically by the new method (for the
string 6, the notes were generated by transposing the string 1 one
Octave UP instead of down). The test data file was updated
accordingly.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>